@@ -1,7 +1,7 @@
 name: record_workspace
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dev_dependencies:
-  melos: ^3.1.1
+  melos: ^3.4.0
