@@ -7,12 +7,13 @@ version: 1.0.0
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   path: ^1.8.0
   path_provider: ^2.0.1
-  audioplayers: ^5.1.0
+  audioplayers: ^5.2.1
 
   flutter:
     sdk: flutter
@@ -21,7 +22,7 @@ dependencies:
     path: ../
 
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.0
 
 # The following section is specific to Flutter.
 flutter:
