@@ -1,22 +1,23 @@
 name: record
-description: Audio recorder from microphone to a given file path with multiple codecs, bit rate and sampling rate options.
-version: 5.0.0-beta.3
+description: Audio recorder from microphone to file or stream with multiple codecs, bit rate and sampling rate options.
+version: 5.0.4
 homepage: https://github.com/llfbandit/record/tree/master/record
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
-  flutter: ">=2.8.1"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  uuid: ^3.0.0
+  # https://pub.dev/packages/uuid
+  uuid: ">=3.0.7 <5.0.0"
 
-  record_platform_interface: ^1.0.1
-  record_web: ^1.0.0-beta.2
-  record_windows: ^1.0.0-beta.2
-  record_linux: '>=0.5.0 < 1.0.0'
+  record_platform_interface: ^1.0.2
+  record_web: ^1.0.4
+  record_windows: ^1.0.2
+  record_linux: '>=0.5.0 <1.0.0'
   record_android:
     git:
       url: https://github.com/wmpsro/record.git
@@ -27,6 +28,7 @@ dependencies:
       url: https://github.com/wmpsro/record.git
       ref: master
       path: record_darwin/
+
 
 dev_dependencies:
   # https://pub.dev/packages/flutter_lints
