@@ -1,17 +1,17 @@
 name: record_android
 description: Android specific implementation for record package called by record_platform_interface.
-version: 1.0.0-beta.2
+version: 1.0.4
 homepage: https://github.com/llfbandit/record/tree/master/record_android
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
-  flutter: ">=2.8.1"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  record_platform_interface: ^1.0.1
+  record_platform_interface: ^1.0.2
 
 flutter:
   plugin:
