@@ -1,10 +1,11 @@
 name: record_platform_interface
 description: A common interface for the record package to call dedicated platforms with method channel.
 homepage: https://github.com/llfbandit/record/tree/master/record_platform_interface
-version: 1.0.1
+version: 1.0.2
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
@@ -13,4 +14,4 @@ dependencies:
   plugin_platform_interface: ^2.0.2
 
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.3
