@@ -1,25 +1,21 @@
 name: record_windows
 description: Windows specific implementation for record package called by record_platform_interface.
-version: 1.0.0-beta.2+1
+version: 1.0.2
 homepage: https://github.com/llfbandit/record/tree/master/record_windows
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
-  flutter: ">=2.8.1"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  record_platform_interface: ^1.0.1
+  record_platform_interface: ^1.0.2
   path: ^1.8.0
 
-# dependency_overrides:
-#   record_platform_interface:
-#     path: ../record_platform_interface
-
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.3
 
 flutter:
   plugin:
