@@ -1,24 +1,10 @@
 package com.llfbandit.record.record
 
 import android.util.Log
-import com.llfbandit.record.record.encoder.EncoderListener
-import com.llfbandit.record.record.encoder.IEncoder
-import com.llfbandit.record.record.format.AacFormat
-import com.llfbandit.record.record.format.AmrNbFormat
-import com.llfbandit.record.record.format.AmrWbFormat
-import com.llfbandit.record.record.format.FlacFormat
-import com.llfbandit.record.record.format.Format
-import com.llfbandit.record.record.format.OpusFormat
-import com.llfbandit.record.record.format.PcmFormat
-import com.llfbandit.record.record.format.WaveFormat
 import com.llfbandit.record.record.stream.RecorderRecordStreamHandler
 import com.llfbandit.record.record.stream.RecorderStateStreamHandler
-import java.io.File
-import java.nio.ByteBuffer
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.atomic.AtomicBoolean
 
-internal interface OnAudioRecordListener {
+interface OnAudioRecordListener {
     fun onRecord()
     fun onPause()
     fun onStop()
@@ -27,30 +13,40 @@ internal interface OnAudioRecordListener {
 }
 
 class AudioRecorder(
-    // Recording config
-    private val config: RecordConfig,
     // Recorder streams
     private val recorderStateStreamHandler: RecorderStateStreamHandler,
     private val recorderRecordStreamHandler: RecorderRecordStreamHandler
 ) : OnAudioRecordListener {
-    private var recorderThread: RecordThread? = null
+    companion object {
+        private val TAG = AudioRecorder::class.java.simpleName
+    }
 
+    // Recorder thread with which we will interact
+    private var recorderThread: RecordThread? = null
     // Amplitude
     private var maxAmplitude = -160.0
+    // Recording config
+    private var config: RecordConfig? = null
+    // Stop callback to be synchronized between stop method return & record stop
+    private var stopCb: ((path: String?) -> Unit)? = null
 
     /**
      * Starts the recording with the given config.
      */
     @Throws(Exception::class)
-    fun start() {
-        recorderThread = RecordThread(config)
+    fun start(config: RecordConfig) {
+        this.config = config
+
+        recorderThread = RecordThread(config, this)
         recorderThread!!.start()
     }
 
     /**
      * Stops the recording.
      */
-    fun stop() {
+    fun stop(stopCb: ((path: String?) -> Unit)?) {
+        this.stopCb = stopCb
+
         recorderThread?.stopRecording()
     }
 
@@ -105,19 +101,23 @@ class AudioRecorder(
     }
 
     fun dispose() {
-        stop()
+        stop(null)
     }
 
+    // OnAudioRecordListener
     override fun onRecord() {
-        updateState(RecordState.RECORD)
+        recorderStateStreamHandler.sendStateEvent(RecordState.RECORD.id)
     }
 
     override fun onPause() {
-        updateState(RecordState.PAUSE)
+        recorderStateStreamHandler.sendStateEvent(RecordState.PAUSE.id)
     }
 
     override fun onStop() {
-        updateState(RecordState.STOP)
+        stopCb?.invoke(config?.path)
+        stopCb = null
+
+        recorderStateStreamHandler.sendStateEvent(RecordState.STOP.id)
     }
 
     override fun onFailure(ex: Exception) {
@@ -125,166 +125,7 @@ class AudioRecorder(
         recorderStateStreamHandler.sendStateErrorEvent(ex)
     }
 
-    private fun updateState(state: RecordState) {
-        recorderStateStreamHandler.sendStateEvent(state.id)
-    }
-
     override fun onAudioChunk(chunk: ByteArray) {
         recorderRecordStreamHandler.sendRecordChunkEvent(chunk)
-    }
-
-    private inner class RecordThread(private val config: RecordConfig) : Thread(), EncoderListener {
-        private var reader: PCMReader? = null
-        private var audioEncoder: IEncoder? = null
-
-        // Signals whether a recording is in progress (true) or not (false).
-        private val isRecording = AtomicBoolean(false)
-
-        // Signals whether a recording is paused (true) or not (false).
-        private val isPaused = AtomicBoolean(false)
-        private var hasBeenCanceled = false
-
-        private val completion = CountDownLatch(1)
-
-        override fun onEncoderDataSize(): Int = reader?.bufferSize ?: 0
-
-        override fun onEncoderDataNeeded(byteBuffer: ByteBuffer): Int {
-            if (isPaused()) return 0
-
-            return reader?.read(byteBuffer) ?: 0
-        }
-
-        override fun onEncoderFailure(ex: Exception) {
-            onFailure(ex)
-        }
-
-        override fun onEncoderStream(bytes: ByteArray) {
-            onAudioChunk(bytes)
-        }
-
-        override fun onEncoderStop() {
-            audioEncoder?.release()
-
-            reader?.stop()
-            reader?.release()
-
-            if (hasBeenCanceled) {
-                deleteFile()
-            }
-
-            updateState(RecordState.STOP)
-
-            completion.countDown()
-        }
-
-        fun isRecording(): Boolean {
-            return audioEncoder != null && isRecording.get()
-        }
-
-        fun isPaused(): Boolean {
-            return audioEncoder != null && isPaused.get()
-        }
-
-        fun pauseRecording() {
-            if (isRecording()) {
-                updateState(RecordState.PAUSE)
-            }
-        }
-
-        fun resumeRecording() {
-            if (isPaused()) {
-                updateState(RecordState.RECORD)
-            }
-        }
-
-        fun stopRecording() {
-            if (isRecording()) {
-                audioEncoder?.stop()
-            }
-        }
-
-        fun cancelRecording() {
-            if (isRecording()) {
-                hasBeenCanceled = true
-                audioEncoder?.stop()
-            } else {
-                deleteFile()
-            }
-        }
-
-        fun getAmplitude(): Double = reader?.getAmplitude() ?: -160.0
-
-        override fun run() {
-            try {
-                val format = selectFormat()
-                val mediaFormat = format.getMediaFormat(config)
-
-                reader = PCMReader(config, mediaFormat)
-
-                audioEncoder = format.getEncoder(config, this)
-
-                reader?.start()
-                audioEncoder?.start()
-
-                updateState(RecordState.RECORD)
-
-                completion.await()
-            } catch (ignored: InterruptedException) {
-            } catch (ex: Exception) {
-                onFailure(ex)
-                onEncoderStop()
-            }
-        }
-
-        private fun selectFormat(): Format {
-            when (config.encoder) {
-                AudioEncoder.aacLc, AudioEncoder.aacEld, AudioEncoder.aacHe -> return AacFormat()
-                AudioEncoder.amrNb -> return AmrNbFormat()
-                AudioEncoder.amrWb -> return AmrWbFormat()
-                AudioEncoder.flac -> return FlacFormat()
-                AudioEncoder.pcm16bits -> return PcmFormat()
-                AudioEncoder.opus -> return OpusFormat()
-                AudioEncoder.wav -> return WaveFormat()
-            }
-            throw Exception("Unknown format: " + config.encoder)
-        }
-
-        private fun updateState(state: RecordState) {
-            when (state) {
-                RecordState.PAUSE -> {
-                    isRecording.set(true)
-                    isPaused.set(true)
-                    onPause()
-                }
-
-                RecordState.RECORD -> {
-                    isRecording.set(true)
-                    isPaused.set(false)
-                    onRecord()
-                }
-
-                RecordState.STOP -> {
-                    isRecording.set(false)
-                    isPaused.set(false)
-                    onStop()
-                }
-            }
-        }
-
-        private fun deleteFile() {
-            val path = config.path
-
-            if (path != null) {
-                val file = File(path)
-
-                if (file.exists()) {
-                    file.delete()
-                }
-            }
-        }
-    }
-
-    companion object {
-        private val TAG = AudioRecorder::class.java.simpleName
     }
 }
