@@ -101,16 +101,13 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
         result(FlutterError(code: "record", message: error.localizedDescription, details: nil))
       }
     case "stop":
-      let path = recorder.stop()
-      result(path)
+      recorder.stop { path in
+        result(path)
+      }
     case "cancel":
       do {
-        if let path = recorder.stop() {
-          try recorder.deleteFile(path: path)
-        }
+        try recorder.cancel()
         result(nil)
-      } catch RecorderError.error(let message, let details) {
-        result(FlutterError(code: "record", message: message, details: details))
       } catch {
         result(FlutterError(code: "record", message: error.localizedDescription, details: nil))
       }
@@ -118,8 +115,12 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
       recorder.pause()
       result(nil)
     case "resume":
-      recorder.resume()
-      result(nil)
+      do {
+        try recorder.resume()
+        result(nil)
+      } catch {
+        result(FlutterError(code: "record", message: error.localizedDescription, details: nil))
+      }
     case "isPaused":
       let isPaused = recorder.isPaused()
       result(isPaused)
@@ -140,13 +141,12 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
       
       result(recorder.isEncoderSupported(encoder))
     case "listInputDevices":
-      let devs = recorder.listInputDevices().map { (device) -> [String : Any] in
-        return [
-          "id": device.uniqueID,
-          "label": device.localizedName,
-        ]
+      do {
+        let devices = try recorder.listInputDevices()
+        result(devices.map({ dev in dev.toMap() }))
+      } catch {
+        result(FlutterError(code: "record", message: error.localizedDescription, details: nil))
       }
-      result(devs)
     case "dispose":
       m_recorders.removeValue(forKey: recorderId)
       recorder.dispose()
@@ -155,7 +155,7 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
       result(FlutterMethodNotImplemented)
     }
   }
-  
+
   private func hasPermission(_ result: @escaping FlutterResult) {
     switch AVCaptureDevice.authorizationStatus(for: .audio) {
     case .authorized:
@@ -180,12 +180,17 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
       return nil
     }
     
+    var device: Device? = nil
+    if let deviceMap = args["device"] as? [String : Any] {
+      device = Device(map: deviceMap)
+    }
+    
     let config = RecordConfig(
       encoder: encoder,
       bitRate: args["bitRate"] as? Int ?? 128000,
       sampleRate: args["sampleRate"] as? Int ?? 44100,
       numChannels: args["numChannels"] as? Int ?? 2,
-      device: args["device"] as? [String : Any],
+      device: device,
       autoGain: args["autoGain"] as? Bool ?? false,
       echoCancel: args["echoCancel"] as? Bool ?? false,
       noiseSuppress: args["noiseSuppress"] as? Bool ?? false
@@ -216,7 +221,7 @@ public class SwiftRecordPlugin: NSObject, FlutterPlugin {
   }
 }
 
-class StateStreamHandler: NSObject, FlutterStreamHandler {
+public class StateStreamHandler: NSObject, FlutterStreamHandler {
   var eventSink: FlutterEventSink?
   
   public func onListen(
@@ -233,7 +238,7 @@ class StateStreamHandler: NSObject, FlutterStreamHandler {
   }
 }
 
-class RecordStreamHandler: NSObject, FlutterStreamHandler {
+public class RecordStreamHandler: NSObject, FlutterStreamHandler {
   var eventSink: FlutterEventSink?
   
   public func onListen(
