@@ -1,13 +1,55 @@
 import AVFoundation
 
-extension Recorder {
-#if os(iOS)
+func listInputs() throws -> [Device] {
+  var devices: [Device] = []
+
+  try listInputDevices()?.forEach { input in
+    devices.append(Device(id: input.uid, label: input.portName))
+  }
+
+  return devices
+}
+
+func listInputDevices() throws -> [AVAudioSessionPortDescription]? {
+  let audioSession = AVAudioSession.sharedInstance()
+  let options: AVAudioSession.CategoryOptions = [.defaultToSpeaker, .allowBluetooth]
+  
+  do {
+    try audioSession.setCategory(.playAndRecord, options: options)
+  } catch {
+    throw RecorderError.error(message: "Failed to list inputs", details: "setCategory: \(error.localizedDescription)")
+  }
+
+  return audioSession.availableInputs
+}
+
+private func setInput(_ config: RecordConfig) throws {
+  guard let device = config.device else {
+    return
+  }
+  
+  let inputs = try listInputDevices()
+  guard let inputs = inputs else {
+    return
+  }
+  
+  let audioSession = AVAudioSession.sharedInstance()
+  
+  for input in inputs {
+    if input.uid == device.id {
+      try audioSession.setPreferredInput(input)
+      break
+    }
+  }
+}
+
+extension AudioRecordingDelegate {
   func initAVAudioSession(config: RecordConfig) throws {
     let audioSession = AVAudioSession.sharedInstance()
-    let options: AVAudioSession.CategoryOptions = [.allowBluetooth]
+    let options: AVAudioSession.CategoryOptions = [.defaultToSpeaker, .allowBluetooth]
     
     do {
-      try audioSession.setCategory(.record, options: options)
+      try audioSession.setCategory(.playAndRecord, options: options)
     } catch {
       throw RecorderError.error(message: "Failed to start recording", details: "setCategory: \(error.localizedDescription)")
     }
@@ -19,17 +61,25 @@ extension Recorder {
     }
     
     do {
-      try audioSession.setActive(true, options: .notifyOthersOnDeactivation) // Must be done before setting channels & buffer duration
+      try audioSession.setActive(true, options: .notifyOthersOnDeactivation) // Must be done before setting channels and others
     } catch {
       throw RecorderError.error(message: "Failed to start recording", details: "setActive: \(error.localizedDescription)")
     }
     
     do {
-      try audioSession.setPreferredInputNumberOfChannels(
-        (config.numChannels > audioSession.inputNumberOfChannels) ? audioSession.inputNumberOfChannels : max(1, config.numChannels)
-      )
+      let newPreferredInputNumberOfChannels = min(config.numChannels, audioSession.maximumInputNumberOfChannels)
+
+      if newPreferredInputNumberOfChannels > 0 {
+        try audioSession.setPreferredInputNumberOfChannels(newPreferredInputNumberOfChannels)
+      }
     } catch {
       throw RecorderError.error(message: "Failed to start recording", details: "setPreferredInputNumberOfChannels: \(error.localizedDescription)")
+    }
+    
+    do {
+      try setInput(config)
+    } catch {
+      throw RecorderError.error(message: "Failed to start recording", details: "setInput: \(error.localizedDescription)")
     }
     
     NotificationCenter.default.addObserver(
@@ -39,26 +89,29 @@ extension Recorder {
       using: onAudioSessionInterruption)
   }
 
-  func onAudioSessionInterruption(notification: Notification) -> Void {
-      guard let userInfo = notification.userInfo,
-            let typeValue = userInfo[AVAudioSessionInterruptionTypeKey] as? UInt,
-            let type = AVAudioSession.InterruptionType(rawValue: typeValue) else {
-        return
-      }
+  private func onAudioSessionInterruption(notification: Notification) -> Void {
+    guard let userInfo = notification.userInfo,
+          let typeValue = userInfo[AVAudioSessionInterruptionTypeKey] as? UInt,
+          let type = AVAudioSession.InterruptionType(rawValue: typeValue) else {
+      return
+    }
+    
+    if type == AVAudioSession.InterruptionType.began {
+      pause()
+    } else if type == AVAudioSession.InterruptionType.ended {
+      guard let optionsValue = userInfo[AVAudioSessionInterruptionOptionKey] as? UInt else { return }
       
-      if type == AVAudioSession.InterruptionType.began {
-        pause()
-      } else if type == AVAudioSession.InterruptionType.ended {
-        guard let optionsValue = userInfo[AVAudioSessionInterruptionOptionKey] as? UInt else { return }
-        
-        let options = AVAudioSession.InterruptionOptions(rawValue: optionsValue)
-
-        if options.contains(.shouldResume) {
-          resume()
-        } else {
-          _ = stop()
+      let options = AVAudioSession.InterruptionOptions(rawValue: optionsValue)
+      
+      if options.contains(.shouldResume) {
+        do {
+          try resume()
+        } catch {
+          stop(completionHandler: {(path) -> () in })
         }
+      } else {
+        stop(completionHandler: {(path) -> () in })
       }
     }
-#endif
+  }
 }
