@@ -1,7 +1,7 @@
 import 'dart:async';
+import 'dart:js_util' as jsu;
 import 'dart:math';
 import 'dart:typed_data';
-import 'dart:js_util' as jsu;
 
 import 'package:record_platform_interface/record_platform_interface.dart';
 import 'package:record_web/encoder/encoder.dart';
@@ -18,7 +18,7 @@ class MicRecorderDelegate extends RecorderDelegate {
   // Media stream get from getUserMedia
   MediaStream? _mediaStream;
   AudioContext? _context;
-  StreamController<Uint8List>? _streamController;
+  StreamController<Uint8List>? _recordStreamCtrl;
   Encoder? _encoder;
   // Amplitude
   double _maxAmplitude = kMinAmplitude;
@@ -27,15 +27,7 @@ class MicRecorderDelegate extends RecorderDelegate {
   MicRecorderDelegate({required this.onStateChanged});
 
   @override
-  Future<void> dispose() async {
-    final context = _context;
-    if (context != null && context.state != AudioContextState.closed) {
-      await context.close();
-      _context = null;
-    }
-
-    return _streamController?.close();
-  }
+  Future<void> dispose() => _reset();
 
   @override
   Future<Amplitude> getAmplitude() async {
@@ -78,31 +70,28 @@ class MicRecorderDelegate extends RecorderDelegate {
 
   @override
   Future<Stream<Uint8List>> startStream(RecordConfig config) async {
-    await _streamController?.close();
+    await _recordStreamCtrl?.close();
     final streamController = StreamController<Uint8List>();
 
     try {
       await _start(config, isStream: true);
     } catch (_) {
-      streamController.close();
+      await streamController.close();
       rethrow;
     }
 
-    _streamController = streamController;
+    _recordStreamCtrl = streamController;
 
     return streamController.stream;
   }
 
   @override
   Future<String?> stop() async {
-    await resetContext(_context, _mediaStream);
-    _mediaStream = null;
-    _context = null;
+    await _reset(resetEncoder: false);
 
     final blob = _encoder?.finish();
+    _encoder?.cleanup();
     _encoder = null;
-    _maxAmplitude = kMinAmplitude;
-    _amplitude = kMinAmplitude;
 
     onStateChanged(RecordState.stop);
 
@@ -111,10 +100,16 @@ class MicRecorderDelegate extends RecorderDelegate {
 
   Future<void> _start(RecordConfig config, {bool isStream = false}) async {
     final mediaStream = await initMediaStream(config);
+    _mediaStream = mediaStream;
 
-    final context = AudioContext(AudioContextOptions(
-      sampleRate: config.sampleRate.toDouble(),
-    ));
+    final constraints = mediaStream.getTracks()[0].getConstraints();
+    bool sampleRateSupported = (constraints.sampleRate != null);
+
+    final context = AudioContext(
+      sampleRateSupported
+          ? AudioContextOptions(sampleRate: config.sampleRate.toDouble())
+          : null,
+    );
 
     final source = context.createMediaStreamSource(mediaStream);
 
@@ -125,7 +120,12 @@ class MicRecorderDelegate extends RecorderDelegate {
     final recorder = AudioWorkletNode(
       context,
       'recorder.worklet',
-      AudioWorkletNodeOptions(numberOfOutputs: config.numChannels),
+      AudioWorkletNodeOptions(
+        parameterData: jsu.jsify({
+          'numChannels': config.numChannels,
+          'sampleRate': config.sampleRate,
+        }),
+      ),
     );
     source.connect(recorder).connect(context.destination);
 
@@ -160,7 +160,7 @@ class MicRecorderDelegate extends RecorderDelegate {
 
   void _onMessage(MessageEvent event) {
     // `data` is a int 16 array containing audio samples
-    final output = event.data;
+    final Int16List output = event.data;
 
     _encoder?.encode(output);
     _updateAmplitude(output);
@@ -168,10 +168,9 @@ class MicRecorderDelegate extends RecorderDelegate {
 
   void _onMessageStream(MessageEvent event) {
     // `data` is a int 16 array containing audio samples
-    final output = event.data;
+    final Int16List output = event.data;
 
-    final data = ByteData.sublistView(output);
-    _streamController?.add(data.buffer.asUint8List());
+    _recordStreamCtrl?.add(output.buffer.asUint8List());
     _updateAmplitude(output);
   }
 
@@ -190,5 +189,22 @@ class MicRecorderDelegate extends RecorderDelegate {
     if (_amplitude > _maxAmplitude) {
       _maxAmplitude = _amplitude;
     }
+  }
+
+  Future<void> _reset({bool resetEncoder = true}) async {
+    await resetContext(_context, _mediaStream);
+    _mediaStream = null;
+    _context = null;
+
+    if (resetEncoder) {
+      _encoder?.cleanup();
+      _encoder = null;
+    }
+
+    _maxAmplitude = kMinAmplitude;
+    _amplitude = kMinAmplitude;
+
+    _recordStreamCtrl?.close();
+    _recordStreamCtrl = null;
   }
 }
